requirements.txt
torchextractor==0.3.0
pillow
kmeans-pytorch